refactor(rag): extract context building logic to contextpkg
Moves Context Building, Symbols, HyDE, Architectural, Impact Analysis
and Formatting functionalities to a new cohesive .
Updates tests and internal dependencies to use the new modular code.
