+++ internal/rag/contextpkg/hyde_test.go
@@ -1,4 +1,4 @@
-package rag
+package contextpkg
 
 import (
 	"testing"
